requirements.txt
flask==2.0.1
requests==2.26.0
gradio
pydantic==1.8.2
python-dotenv==0.19.0
google-generativeai==0.1.0
openai==0.27.0
flask-cors==3.0.10
flask-sqlalchemy==2.5.1
psycopg2-binary==2.9.1